requirements.txt
Django==5.0
djangorestframework==3.15.2
gunicorn==20.1.0
celery[redis]==5.4.0
django-redis==5.0.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
drf-yasg==1.21.7
bandit==1.7.9
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.1
cryptography==43.0.0
locust==2.31.3